% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastictranscoder_operations.R
\name{elastictranscoder_read_preset}
\alias{elastictranscoder_read_preset}
\title{The ReadPreset operation gets detailed information about a preset}
\usage{
elastictranscoder_read_preset(Id)
}
\arguments{
\item{Id}{[required] The identifier of the preset for which you want to get detailed
information.}
}
\description{
The ReadPreset operation gets detailed information about a preset.
}
\section{Request syntax}{
\preformatted{elastictranscoder$read_preset(
  Id = "string"
)
}
}

\keyword{internal}
